possibly finished 1

+++ assign0/assign0/da.c
@@ -71,7 +71,7 @@ removeDA(DA *items)
 
 	int size = items->size;
 
-	if (( 0.25 > items->size / items->capacity) && items->capacity != 1)
+	if (( 0.25 > items->size / (double) items->capacity) && items->capacity != 1)
 	{
 		items->capacity = items->capacity / items->factor;
 		items->store = realloc(items->store, sizeof(void *) * items->capacity);
